Get the dataset associated with a ticket_id

Getting a dataset is the second stage in analyzing a simulation from causalbenchmarks.org. Once you have received a ticket_id from get_new_ticket , you can use this function to download the dataset to analyze. Requesting a dataset stars a timer that runs until the results of the analysis are submitted, in order to track the speed of your algorithm.

Usage

get_dataset(ticket_id)

Arguments

ticket_id

a 32-character ticket_id supplied by get_ticket. Please note that this is a secret value and should not be shared with anyone else.

Value

a dataframe containing the data associated with this simulation